Fix: add more types recognition, fix several bugs in global vars initialization -move symbols type and value definition to another class -fix keyErrors -fix 16-digit numbers initialization -add bool recognition -does not change char * and char[N] types for now
